fix(cad): give join's junction knots the dtype of the knots they join
Joining two float32 B-splines raised

    ValueError: The control points must have the same dtype as the B-spline
    space. Got float32 control points and float64 B-spline space.

even when the two pieces met bitwise exactly, so the failure had nothing to do
with the geometry. _concatenate_along_axis built the p junction knots with
np.full(p, u_junction), and u_junction is a Python float, so that array is
float64. Concatenating it with the two float32 knot vectors promoted the merged
vector to float64 while the control points stayed float32, and BsplineSpace1D
refuses that pair. Taking the dtype from the incoming knots fixes it in one
place.

float32 is a supported precision throughout pantr's B-spline layer, so this is
an ordinary float32 model failing to assemble, not an exotic input. A mixed
float32/float64 join was unaffected and stays unaffected: numpy's own promotion
already carried both the knots and the control points to float64 there.

Verified failing on the parent of this branch, and the three regression tests
also cover the paths the fix relies on, since it is only correct if they too
preserve float32: elevate_degree on the join axis (which at degree > 1 also
routes through _try_remove_junction_knots) and insert_knots. Nothing in the file
joined float32 inputs before.

This is unrelated to the shared-boundary check this branch is for; it was found
while testing that check across precisions and is fixed here because it sits in
the function that check is being added to. The 15 tests pinned red by the
previous commit stay red until the next one, which is the fix they are the
target of.

+++ src/pantr/cad/_join.py
@@ -123,11 +123,13 @@ def _concatenate_along_axis(
 
     new_cp = np.concatenate([cp_left, cp_mid_expanded, cp_right], axis=axis)
 
-    # Merge knot vectors
+    # Merge knot vectors.  The junction knots take the incoming knots' dtype: built from a
+    # Python float they would be float64, and concatenating would promote a float32 knot
+    # vector while the control points stayed float32, which BsplineSpace1D refuses.
     u_junction = float(b1.space.spaces[axis].domain[1])
     kl = b1.space.spaces[axis].knots[: -p - 1]
     kr = b2.space.spaces[axis].knots[p + 1 :]
-    kc = np.full(p, u_junction)
+    kc = np.full(p, u_junction, dtype=kl.dtype)
     new_knots = np.concatenate([kl, kc, kr])
 
     spaces = list(b1.space.spaces)
